Moving policy as a subpackage of http package (to accomodate any future protocol based policies (tcp e.g.)
