+++ azure-client-runtime/src/main/java/com/microsoft/azure/v3/AzureProxy.java
@@ -18,14 +18,14 @@ import com.microsoft.azure.v3.credentials.AsyncServiceClientCredentials;
 import com.microsoft.rest.v3.credentials.ServiceClientCredentials;
 import com.microsoft.rest.v3.http.HttpMethod;
 import com.microsoft.rest.v3.http.HttpPipeline;
-import com.microsoft.rest.v3.policy.HttpPipelinePolicy;
+import com.microsoft.rest.v3.http.policy.HttpPipelinePolicy;
 import com.microsoft.rest.v3.http.HttpRequest;
 import com.microsoft.rest.v3.http.HttpResponse;
-import com.microsoft.rest.v3.policy.CookiePolicy;
-import com.microsoft.rest.v3.policy.CredentialsPolicy;
-import com.microsoft.rest.v3.policy.DecodingPolicy;
-import com.microsoft.rest.v3.policy.RetryPolicy;
-import com.microsoft.rest.v3.policy.UserAgentPolicy;
+import com.microsoft.rest.v3.http.policy.CookiePolicy;
+import com.microsoft.rest.v3.http.policy.CredentialsPolicy;
+import com.microsoft.rest.v3.http.policy.DecodingPolicy;
+import com.microsoft.rest.v3.http.policy.RetryPolicy;
+import com.microsoft.rest.v3.http.policy.UserAgentPolicy;
 import com.microsoft.rest.v3.protocol.SerializerAdapter;
 import com.microsoft.rest.v3.protocol.SerializerEncoding;
 import com.microsoft.rest.v3.util.TypeUtil;
