+++ azure-client-runtime/src/test/java/com/microsoft/azure/v3/AzureProxyTests.java
@@ -11,8 +11,8 @@ import com.microsoft.rest.v3.RestException;
 import com.microsoft.rest.v3.http.HttpPipeline;
 import com.microsoft.rest.v3.http.HttpRequest;
 import com.microsoft.rest.v3.http.HttpResponse;
-import com.microsoft.rest.v3.policy.DecodingPolicy;
-import com.microsoft.rest.v3.policy.HttpPipelineOptions;
+import com.microsoft.rest.v3.http.policy.DecodingPolicy;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
 import com.microsoft.rest.v3.protocol.SerializerAdapter;
 import com.microsoft.rest.v3.serializer.JacksonAdapter;
 import com.microsoft.rest.v3.InvalidReturnTypeException;
