+++ azure-client-runtime/src/test/java/com/microsoft/azure/v3/AzureProxyToRestProxyTests.java
@@ -5,8 +5,8 @@ import com.microsoft.rest.v3.http.ContentType;
 import com.microsoft.rest.v3.RestException;
 import com.microsoft.rest.v3.RestResponse;
 import com.microsoft.rest.v3.http.HttpPipeline;
-import com.microsoft.rest.v3.policy.DecodingPolicy;
-import com.microsoft.rest.v3.policy.HttpPipelineOptions;
+import com.microsoft.rest.v3.http.policy.DecodingPolicy;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
 import com.microsoft.rest.v3.protocol.SerializerAdapter;
 import com.microsoft.rest.v3.serializer.JacksonAdapter;
 import com.microsoft.rest.v3.annotations.BodyParam;
