+++ client-runtime/src/main/java/com/microsoft/rest/v3/RestProxy.java
@@ -14,16 +14,16 @@ import com.microsoft.rest.v3.http.HttpHeader;
 import com.microsoft.rest.v3.http.HttpHeaders;
 import com.microsoft.rest.v3.http.HttpMethod;
 import com.microsoft.rest.v3.http.HttpPipeline;
-import com.microsoft.rest.v3.policy.HttpPipelinePolicy;
+import com.microsoft.rest.v3.http.policy.HttpPipelinePolicy;
 import com.microsoft.rest.v3.http.HttpRequest;
 import com.microsoft.rest.v3.http.HttpResponse;
 import com.microsoft.rest.v3.http.UrlBuilder;
-import com.microsoft.rest.v3.policy.CookiePolicy;
-import com.microsoft.rest.v3.policy.CredentialsPolicy;
-import com.microsoft.rest.v3.policy.DecodingPolicy;
-import com.microsoft.rest.v3.policy.HttpPipelineOptions;
-import com.microsoft.rest.v3.policy.RetryPolicy;
-import com.microsoft.rest.v3.policy.UserAgentPolicy;
+import com.microsoft.rest.v3.http.policy.CookiePolicy;
+import com.microsoft.rest.v3.http.policy.CredentialsPolicy;
+import com.microsoft.rest.v3.http.policy.DecodingPolicy;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
+import com.microsoft.rest.v3.http.policy.RetryPolicy;
+import com.microsoft.rest.v3.http.policy.UserAgentPolicy;
 import com.microsoft.rest.v3.protocol.HttpResponseDecoder;
 import com.microsoft.rest.v3.protocol.SerializerAdapter;
 import com.microsoft.rest.v3.protocol.SerializerEncoding;
