+++ client-runtime/src/main/java/com/microsoft/rest/v3/http/HttpPipeline.java
@@ -6,8 +6,7 @@
 
 package com.microsoft.rest.v3.http;
 
-import com.microsoft.rest.v3.policy.HttpPipelinePolicy;
-import com.microsoft.rest.v3.policy.HttpPipelineOptions;
+import com.microsoft.rest.v3.http.policy.HttpPipelinePolicy;
 import reactor.core.publisher.Mono;
 
 import java.util.Objects;
