+++ client-runtime/src/main/java/com/microsoft/rest/v3/http/policy/CredentialsPolicy.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
 
 import com.microsoft.rest.v3.credentials.ServiceClientCredentials;
 import com.microsoft.rest.v3.http.HttpPipelineCallContext;
