+++ client-runtime/src/main/java/com/microsoft/rest/v3/http/policy/HttpLogDetailLevel.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
 
 /**
  * The level of detail to log on HTTP messages.
