+++ client-runtime/src/main/java/com/microsoft/rest/v3/http/policy/PortPolicy.java
@@ -4,10 +4,11 @@
  * license information.
  */
 
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
 
 import com.microsoft.rest.v3.http.HttpPipelineCallContext;
 import com.microsoft.rest.v3.http.HttpPipelineLogLevel;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
 import com.microsoft.rest.v3.http.HttpResponse;
 import com.microsoft.rest.v3.http.NextPolicy;
 import com.microsoft.rest.v3.http.UrlBuilder;
@@ -16,43 +17,43 @@ import reactor.core.publisher.Mono;
 import java.net.MalformedURLException;
 
 /**
- * The Pipeline policy that adds a given protocol to each HttpRequest.
+ * The Pipeline policy that adds a given port to each HttpRequest.
  */
-public class ProtocolPolicy extends AbstractPipelinePolicy {
-    private final String protocol;
+public class PortPolicy extends AbstractPipelinePolicy {
+    private final int port;
     private final boolean overwrite;
 
     /**
-     * Create a new ProtocolPolicy.
+     * Create a new PortPolicy object.
      *
-     * @param protocol The protocol to set on every HttpRequest.
+     * @param port The port to set on every HttpRequest.
      */
-    public ProtocolPolicy(String protocol) {
-        this(protocol, true, new HttpPipelineOptions(null));
+    public PortPolicy(int port) {
+        this(port, true, new HttpPipelineOptions(null));
     }
 
     /**
-     * Create a new ProtocolPolicy.
+     * Create a new PortPolicy object.
      *
-     * @param protocol The protocol to set.
-     * @param overwrite Whether or not to overwrite a HttpRequest's protocol if it already has one.
+     * @param port The port to set.
+     * @param overwrite Whether or not to overwrite a HttpRequest's port if it already has one.
      * @param options the request options
      */
-    public ProtocolPolicy(String protocol, boolean overwrite, HttpPipelineOptions options) {
+    public PortPolicy(int port, boolean overwrite, HttpPipelineOptions options) {
         super(options);
-        this.protocol = protocol;
+        this.port = port;
         this.overwrite = overwrite;
     }
 
     @Override
     public Mono<HttpResponse> process(HttpPipelineCallContext context, NextPolicy next) {
         final UrlBuilder urlBuilder = UrlBuilder.parse(context.httpRequest().url());
-        if (overwrite || urlBuilder.scheme() == null) {
+        if (overwrite || urlBuilder.port() == null) {
             if (shouldLog(HttpPipelineLogLevel.INFO)) {
-                log(HttpPipelineLogLevel.INFO, "Setting protocol to {0}", protocol);
+                log(HttpPipelineLogLevel.INFO, "Changing port to {0}", port);
             }
             try {
-                context.httpRequest().withUrl(urlBuilder.withScheme(protocol).toURL());
+                context.httpRequest().withUrl(urlBuilder.withPort(port).toURL());
             } catch (MalformedURLException e) {
                 return Mono.error(e);
             }
