+++ client-runtime/src/main/java/com/microsoft/rest/v3/http/policy/RetryPolicy.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
 
 import com.microsoft.rest.v3.http.HttpPipelineCallContext;
 import com.microsoft.rest.v3.http.HttpRequest;
