+++ client-runtime/src/main/java/com/microsoft/rest/v3/http/policy/package-info.java
@@ -2,5 +2,5 @@
  * Code related to creating proxy implementations of interfaces that have been generated from
  * Swagger specifications.
  */
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
