+++ client-runtime/src/test/java/com/microsoft/rest/v3/RestProxyStressTests.java
@@ -18,15 +18,15 @@ import com.microsoft.rest.v3.http.ContentType;
 import com.microsoft.rest.v3.http.HttpHeaders;
 import com.microsoft.rest.v3.http.HttpPipeline;
 import com.microsoft.rest.v3.http.HttpPipelineCallContext;
-import com.microsoft.rest.v3.policy.HttpLoggingPolicy;
-import com.microsoft.rest.v3.policy.HttpPipelinePolicy;
+import com.microsoft.rest.v3.http.policy.HttpLoggingPolicy;
+import com.microsoft.rest.v3.http.policy.HttpPipelinePolicy;
 import com.microsoft.rest.v3.http.HttpResponse;
 import com.microsoft.rest.v3.http.NextPolicy;
-import com.microsoft.rest.v3.policy.AddDatePolicy;
-import com.microsoft.rest.v3.policy.AddHeadersPolicy;
-import com.microsoft.rest.v3.policy.HostPolicy;
-import com.microsoft.rest.v3.policy.HttpLogDetailLevel;
-import com.microsoft.rest.v3.policy.HttpPipelineOptions;
+import com.microsoft.rest.v3.http.policy.AddDatePolicy;
+import com.microsoft.rest.v3.http.policy.AddHeadersPolicy;
+import com.microsoft.rest.v3.http.policy.HostPolicy;
+import com.microsoft.rest.v3.http.policy.HttpLogDetailLevel;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
 import com.microsoft.rest.v3.util.FluxUtil;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
