+++ client-runtime/src/test/java/com/microsoft/rest/v3/RestProxyWithMockTests.java
@@ -13,7 +13,7 @@ import com.microsoft.rest.v3.http.HttpRequest;
 import com.microsoft.rest.v3.http.HttpResponse;
 import com.microsoft.rest.v3.http.MockHttpClient;
 import com.microsoft.rest.v3.http.MockHttpResponse;
-import com.microsoft.rest.v3.policy.HttpPipelineOptions;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
 import org.junit.Test;
 import reactor.core.publisher.Mono;
