+++ client-runtime/src/test/java/com/microsoft/rest/v3/http/policy/HttpPipelineOptionsTests.java
@@ -1,6 +1,7 @@
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
 
 import com.microsoft.rest.v3.http.HttpPipelineLogLevel;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
 import com.microsoft.rest.v3.http.MockHttpPipelineLogger;
 import org.junit.Test;
