+++ client-runtime/src/test/java/com/microsoft/rest/v3/http/policy/ProxyAuthenticationPolicyTests.java
@@ -1,7 +1,8 @@
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
 
 import com.microsoft.rest.v3.http.HttpMethod;
 import com.microsoft.rest.v3.http.HttpPipeline;
+import com.microsoft.rest.v3.http.HttpPipelineOptions;
 import com.microsoft.rest.v3.http.HttpRequest;
 import com.microsoft.rest.v3.http.MockHttpClient;
 import org.junit.Test;
