+++ client-runtime/src/test/java/com/microsoft/rest/v3/http/policy/RetryPolicyTests.java
@@ -4,7 +4,7 @@
  * license information.
  */
 
-package com.microsoft.rest.v3.policy;
+package com.microsoft.rest.v3.http.policy;
 
 import com.microsoft.rest.v3.http.*;
 import org.junit.Assert;
